Updated Manhattan plots
Almost finished with the Manhattan plots, need only to fix the colour for the
BW option. This should not take long. Also finished the PCA plot with a new
MDS option. Still no option to do the analysis on (most) the most variable
probes.

Also added a draft vignette, but it is still empty. Need to add some
information there. One for vissualisation and QC and one for high performance
GWAS analysis.